Create control Thorlabs ELL6 2-position slider
In Nijmegen, we created this class to control a Thorlabs linear 2-position slider within the MIC. This class controls an Elliptec ELL6 shutter, which is USB connected via a rs232-to-USB2.0 board.  The shutter and the board are delivered as a package, see the Thorlabs catalog, # ELL6K. The code is adapted from ShutterTLL, which was written by Farzin. 

With help by Sandeep and Keith, we would like to implement other classes as well, such that we can control our home built multi-color 3D storm setup in Nijmegen, the Netherlands. A nano stage from PI, a thorlabs filterwheel and Coherent OPSL and OBIS lasers will be added soon.

Best regards,
Gert-Jan 
gert-jan.bakker<at>radboudumc.nl